+++ src/alerts/alerts.provider.jsx
@@ -1,35 +1,45 @@
 /* eslint-disable import/prefer-default-export */
-
 import React from 'react';
-import { createStore,
-         applyMiddleware,
-         compose,
-         combineReducers } from 'redux';
+import PropTypes from 'prop-types';
+import {
+  createStore,
+  applyMiddleware,
+  combineReducers,
+} from 'redux';
 import { Provider } from 'react-redux';
-import { IntlProvider, intlReducer } from 'react-intl-redux';
 import thunk from 'redux-thunk';
+import { composeWithDevTools } from 'redux-devtools-extension/developmentOnly';
 import { alertsReducer } from './alerts.reducer';
 import { OCAlert } from './alerts.actions';
 import { OCAlerts } from './alerts.component';
 
 import './alerts.scss';
 
+const composeEnhancers = composeWithDevTools({ name: 'OCAlerts' });
 const store = createStore(
   combineReducers({
-    alertsReducer,
-    intl: intlReducer,
+    alerts: alertsReducer,
   }),
-  compose(applyMiddleware(thunk)),
+  undefined,
+  composeEnhancers(applyMiddleware(thunk)),
 );
 
 OCAlert.setStore(store);
 
-export const OCAlertsProvider = function OCAlertsProvider() {
-  return (
-    <Provider store={store}>
-      <IntlProvider>
-        <OCAlerts />
-      </IntlProvider>
-    </Provider>
-  );
-};
+export class OCAlertsProvider extends React.PureComponent {
+  static propTypes = {
+    containerStyle: PropTypes.object, // eslint-disable-line
+  }
+
+  static defaultProps = {
+    containerStyle: {},
+  }
+
+  render() {
+    return (
+      <Provider store={store}>
+        <OCAlerts containerStyle={this.props.containerStyle} />
+      </Provider>
+    );
+  }
+}